pyproject.toml
[project]
name = "TASKHUB"
version = "1.0.0"
description = "API de taskhub app"
authors = [
    {name = "Oyhs-co",email = "oyhsotelo@gmail.com"}
]
license = {text = "LICENSE"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "supabase (>=2.25.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
